Added AVFoundation framework to build phase links. Added audiosession to ensure podcasts play when NHIC app goes to the background. Also cleared an analyzer warning in the announceDetailViewController. Changed tableview cells from detail disclosure (i)'s to just disclosure chevrons. Added deselect row commands to home screen and announcement screen.

+++ nhicapp/AnnounceDetailViewController.m
@@ -75,7 +75,7 @@
     if(length <= 35)
         self.labelView.numberOfLines = 1;
     else {
-        NSLog(@"length is: %d", length);
+        NSLog(@"length is: %lu", (unsigned long)length);
     
     self.labelView.numberOfLines = length % 22;
     }
